+++ packages/format-message-formats/index.js
@@ -16,6 +16,9 @@ module.exports = {
     },
     percent: {
       style: 'percent'
+    },
+    default: {
+      style: 'decimal'
     }
   },
   date: {
@@ -39,6 +42,11 @@ module.exports = {
       day: 'numeric',
       year: 'numeric',
       weekday: 'long'
+    },
+    default: {
+      month: 'short',
+      day: 'numeric',
+      year: 'numeric'
     }
   },
   time: {
@@ -62,6 +70,11 @@ module.exports = {
       minute: 'numeric',
       second: 'numeric',
       timeZoneName: 'short'
+    },
+    default: {
+      hour: 'numeric',
+      minute: 'numeric',
+      second: 'numeric'
     }
   }
 }