+++ packages/babel-plugin-transform-format-message/CHANGELOG.md
@@ -5,6 +5,7 @@
 **New Feature**
   * Transform messages from JSX marked with translate="yes"
   * Allow `missingReplacement` to be a function
+  * Remove format-message import if all uses were inlined
 
 ## 4.1.0